refactor : BonusBall 클래스 및 테스트 코드 삭제, BonusBall에 있던 로직을 LottoNumber에 추가, bonusBall 의 네이밍을 bonusNumber로 수정,
